Rename sans_fitter.py to fitter.py
Removes the package-name echo (sans_fitter.sans_fitter.SANSFitter).
Public API is unchanged; SANSFitter and get_all_models are still
re-exported from the package root.
