+++ src/sans_fitter/__init__.py
@@ -7,7 +7,7 @@ from .parameter_manager import ParameterManager
 from .polydispersity import PD_DEFAULTS, PD_DISTRIBUTION_TYPES
 from .pr_inversion import InsufficientDataError, PrEstimationError, PrResult
 from .results import FitResultContract, PosteriorSummary
-from .sans_fitter import SANSFitter, get_all_models
+from .fitter import SANSFitter, get_all_models
 
 __version__ = '0.3.0'
 __all__ = [
